lvl up function

+++ src/main/java/isep/rpg/SpellCaster.java
@@ -10,16 +10,16 @@ public abstract class SpellCaster extends Hero {
     public int getManaPoints(){
         return this.manaPoints;
     }
-    public void receiveMana(int mana){
+    public void receiveMana(float mana){
         this.manaPoints+=mana;
     }
 
 
     public void useFood(Potion  pot) {
-        int heal = pot.gainHealth()*this.getEffect();
+        float heal = pot.gainHealth()*this.getEffect();
         this.healsPotion(heal);
 
-        int mana = pot.gainMana()*this.getEffect();
+        float mana = pot.gainMana()*this.getEffect();
         this.receiveMana(mana);
     }
